生成kml

+++ 后端/src/main/java/com/cy/demo1/data/Data2.java
@@ -1,7 +1,7 @@
 package com.cy.demo1.data;
 
 public class Data2 {
-    public int[] x;
-    public int[] y;
+    public double[] x;
+    public double[] y;
     public int[] path;
 }
